create sidebar, begin working on more views
this removes the selection box from the accounts page. the
accounts page is now blank as is the account details page.
there is a new transactions page that doesn't do anything
yet. one of the big refactors was to have the routes
namespace  each of the route types.

i think there is a lack of centralized code for adding new
views. adding the transactions route, for example, meant
modifying several different namespaces. perhaps this is
unavoidable, but I think it could be solved by having a
route -> view mapping in the routes namespace.

another change in here is to remove the ^:figheel-always
metadata from every namespace. I think it's overkill /
should be used when there is actually a known reason.
the figwheel docs indicate that normally a namespace is
loaded every time it changes anyway.

this also meant giving the dev namespace metadata to
tell figwheel not to reload the namespace. this is so
that the database is not refreshed on every save.