so many changes
A lot of this was removing hard-coded keywords and instead referencing
namespace definitions. For example, the db namespace has some defs now.
That namespace also has helper functions which will abstract some of the
work the handlers namespace is doing.

Primarily this commit is moving things in the direction of getting the
transactions to not be loaded during initialization. Instead they will
be queried when they are needed and injected into the main list of
transactions.

I thought a bit about not having one main list of transactions and
instead just having lists for each account. But this breaks down if you
also want to do queries like, for example, show transactions across all
accounts. Or should transactions based on the payee. Having a main list
allows for those queries within the app without duplicating transaction
data all over the place, which would primarily be ugly in terms of
updating the db when a transaction changes.