lots of work to get basic notification showing when waiting for HTTP
As part of this change, I stopped using reagent-forms (at least in the
transaction editor, which I think is the only place using it). It really
does work very well for anything but extremely basic cases. I was
running into an issue with code like this:

```
[:div {:class (when (:property @form) "myclass")}]
```

With that div inside a reagent-form, the `when` only is executed at the
beginning. Changes to `@form` are not respected. This plus my other
experiences leave me with a bad impression of the library and I'd rather
not use it.