get rid of the initialization queries, create a generic HTTP query system
this introduces a more generic way to make an HTTP query
and wait for the results. a query-id is used to uniquely
identify the query, rather than using the parameters to
the handler. you can then pass arbitrary arguments to the
handler, which are forwarded to the HTTP request.

this eliminates the need for duplicating any HTTP request/
response handlers and gets everything automatic handling of
errors.

I also added the i18n file which was missing from the
previous commit.