CSS fixes for navbar on mobile
The hamburger button was a) not working b) UX fail. This changes the
mobile version to just show the same navbar as on the larger screen, but
excludes "Buckit" from the navbar.

This also adds a transactions routes (which is not implemented) but
provides a placeholder to get to transactions on mobile. The next change
would be to add a dropdown that let's you select the account and to show
all transactions in the global "transactions" route.